Add initial version of vscode nvim init.vim
This file is used to configure the nvim in vscode. A lot less plugins are used
because vscode provides its own features.

It works quite well, however, there are still some open things:
  * Git integration mainly missing ([c,]c,gp,gd are ok, but :G is missing)
  * Search-in-files ripgrep integration missing (but vscode is strong there)
  * Navigation through quickfix and other lists missing (what is needed?)
  * Not sure about macros, especially bufdo (not tried yet)
  * Other special features (but what is needed a the end of the day?)

Vscode has the benefit of proper Copilot integration and modern visual features.
This is the main reason why this file was created and why I am playing with
Vscode instead of plain vim. So far I am confident of having done the switch,
however, there are small features that I still miss.